refactor : 코드 리뷰 수정 사항 반영
